+++ vcxproj2cmake.Tests/ConverterTests/LinkerOptionsTests.cs
@@ -1,0 +1,70 @@
+using Microsoft.Extensions.Logging.Abstractions;
+using System.IO.Abstractions.TestingHelpers;
+using Xunit;
+
+namespace vcxproj2cmake.Tests;
+
+public partial class ConverterTests
+{
+    public class LinkerOptionsTests
+    {
+        [Fact]
+        public void Given_AdditionalOptions_When_Converted_Then_OptionsAreWritten()
+        {
+            // Arrange
+            var fileSystem = new MockFileSystem();
+            fileSystem.Directory.SetCurrentDirectory(Environment.CurrentDirectory);
+
+            fileSystem.AddFile(@"Project.vcxproj", new(TestData.Project()
+                .WithLinkSetting("AdditionalOptions", "foo bar")
+                .Build()));
+
+            var converter = new Converter(fileSystem, NullLogger.Instance);
+
+            // Act
+            converter.Convert(
+                projectFiles: [new(@"Project.vcxproj")]);
+
+            // Assert
+            var cmake = fileSystem.GetFile(@"CMakeLists.txt").TextContents;
+
+            Assert.Contains("""
+                target_link_options(Project
+                    PRIVATE
+                        foo
+                        bar
+                )
+                """.TrimEnd(), cmake);
+        }
+
+        [Fact]
+        public void Given_AdditionalOptionsAndPortableMode_When_Converted_Then_OptionsAreGuarded()
+        {
+            // Arrange
+            var fileSystem = new MockFileSystem();
+            fileSystem.Directory.SetCurrentDirectory(Environment.CurrentDirectory);
+
+            fileSystem.AddFile(@"Project.vcxproj", new(TestData.Project()
+                .WithLinkSetting("AdditionalOptions", "foo bar")
+                .Build()));
+
+            var converter = new Converter(fileSystem, NullLogger.Instance);
+
+            // Act
+            converter.Convert(
+                projectFiles: [new(@"Project.vcxproj")],
+                portable: true);
+
+            // Assert
+            var cmake = fileSystem.GetFile(@"CMakeLists.txt").TextContents;
+
+            Assert.Contains("""
+                target_link_options(Project
+                    PRIVATE
+                        $<$<CXX_COMPILER_ID:MSVC>:foo>
+                        $<$<CXX_COMPILER_ID:MSVC>:bar>
+                )
+                """.TrimEnd(), cmake);
+        }
+    }
+}